+++ src/index.ts
@@ -2,6 +2,9 @@
  * @fileoverview RessurectIT specific ESlint rules and configs
  * @author kukjevov
  */
+import eslint from '@eslint/js';
+import stylistic from '@stylistic/eslint-plugin'
+import tseslint from 'typescript-eslint';
 
 import meta from './meta/index.js';
 import rules from './rules/index.js';
@@ -17,44 +20,72 @@ Object.assign(plugin.configs,
 {
     recommended:
     [
+        eslint.configs.recommended,
+        ...tseslint.configs.recommended,
+        stylistic.configs['recommended-flat'],
         {
             plugins:
             {
-                ressurectit: plugin
+                ressurectit: plugin,
             },
             rules:
             {
-                "semi": 
+                'indent': 'off',
+                'semi': 'off',
+                'brace-style': 'off',
+                'linebreak-style': 'off',
+                'quotes': 'off',
+                '@stylistic/semi':
                 [
-                    "error",
-                    "always"
+                    'error',
+                    'always',
                 ],
-                "brace-style": 
+                '@stylistic/brace-style':
                 [
-                    "error",
-                    "allman",
+                    'error',
+                    'allman',
                     {
-                        "allowSingleLine": true
+                        'allowSingleLine': true,
+                    },
+                ],
+                '@stylistic/linebreak-style':
+                [
+                    'error',
+                    'unix',
+                ],
+                '@stylistic/quotes':
+                [
+                    'error',
+                    'single',
+                ],
+                '@stylistic/indent': 
+                [
+                    'warn',
+                    4,
+                    {
+                        'VariableDeclarator': 'first',
+                        'FunctionDeclaration':
+                        {
+                            'parameters': 'first'
+                        },
+                        'FunctionExpression':
+                        {
+                            'parameters': 'first'
+                        },
+                        'CallExpression':
+                        {
+                            'arguments': 'first'
+                        },
+                        'ArrayExpression': 'first',
+                        'ImportDeclaration': 'first'
                     }
                 ],
-                "indent":
-                [
-                    "error",
-                    4
-                ],
-                "linebreak-style":
-                [
-                    "error",
-                    "unix"
-                ],
-                "quotes":
-                [
-                    "error",
-                    "single"
-                ],
-                "@typescript-eslint/no-empty-function": "off",
-                "ressurectit/imports-order": "error",
-                "ressurectit/imports-spacing": "warn"
+                '@typescript-eslint/no-empty-function': 'off',
+                '@typescript-eslint/no-unused-vars': 'off',
+                '@typescript-eslint/no-inferrable-types': 'off',
+                '@typescript-eslint/ban-ts-comment': 'off',
+                'ressurectit/imports-order': 'error',
+                'ressurectit/imports-spacing': 'warn',
             },
         }
     ],