+++ changelog.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Version 0.1.0 (2025-01-23)
+
+### Features
+
+- now usinig new *flat configuration*
+- recommended configuration contains also configuration for `eslint`, `@typescript-eslint` and `@stylistic`
+
+### BREAKING CHANGES
+
+- minimal supported version of `eslint` is `9.18.0`
+- minimal supported version of `@typescript-eslint` is `8.21.0`
+
 ## Version 0.1.0 (2021-06-25)
 
 ### Features